Had some issues. Lesson 10.
Had some issues moving arond folders and getting things to save there,
but i now have it. I also had some issues with the img, but it came down
to just finding the file. I did not link it far enough back so that it
could find it's way forward. :) all is well.